API Key based https REST Routes

+++ src/main/java/com/example/MonaServer/Controller/RestControllerMona.java
@@ -3,17 +3,20 @@ package com.example.MonaServer.Controller;
 import com.example.MonaServer.Entities.Mona;
 import com.example.MonaServer.Entities.Pin;
 import com.example.MonaServer.Entities.StickerType;
+import com.example.MonaServer.Helper.Config;
 import com.example.MonaServer.Repository.*;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
 import java.io.*;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 public class RestControllerMona {
@@ -32,13 +35,22 @@ public class RestControllerMona {
     @Autowired
     UserRepo userRepo;
 
-    @GetMapping(value = "/monas/")
+    @Value("${AUTH_TOKEN_ADMIN}")
+    private String principalRequestValueAdmin;
+
+    @GetMapping(value = "/api/monas/")
     public List<Mona> getMonas () {
         return (List<Mona>) monaRepo.findAll();
     }
 
-    @PostMapping(value = "/monas/")
+    @PostMapping(value = "/api/monas/")
     public void addNewPinToUser(@RequestBody ObjectNode json) throws Exception {
+        if (!json.has("image")) throw new Exception("Error: Field 'image' was not given in request");
+        if (!json.has("latitude")) throw new Exception("Error: Field 'latitude' was not given in request");
+        if (!json.has("longitude")) throw new Exception("Error: Field 'longitude' was not given in request");
+        if (!json.has("username")) throw new Exception("Error: Field 'username' was not given in request");
+        if (!json.has("typeId")) throw new Exception("Error: Field 'typeId' was not given in request");
+        if (!json.has("creationDate")) throw new Exception("Error: Field 'creationDate' was not given in request");
         ObjectMapper mapper = new ObjectMapper();
         ObjectReader reader = mapper.readerFor(new TypeReference<byte[]>() {});
         byte[] image = reader.readValue(json.get("image"));
@@ -52,13 +64,14 @@ public class RestControllerMona {
         }
     }
 
-    @GetMapping(value = "/monas/{pinId}")
+    @GetMapping(value = "/api/monas/{pinId}")
     public Mona getMonaByPinId (@PathVariable("pinId") Long id) {
         return monaRepo.findMonaByPin(pinRepo.findByPinId(id));
     }
 
-    @PutMapping(value = "/monas/{pinId}/")
-    public void updatePictureOfMona(@PathVariable("pinId") Long id, @RequestBody ObjectNode json) throws IOException {
+    @PutMapping(value = "/api/monas/{pinId}/")
+    public void updatePictureOfMona(@PathVariable("pinId") Long id, @RequestBody ObjectNode json) throws Exception {
+        if (!json.has("image")) throw new Exception("Error: Field 'image' was not given in request");
         ObjectMapper mapper = new ObjectMapper();
         ObjectReader reader = mapper.readerFor(new TypeReference<byte[]>() {});
         byte[] image = reader.readValue(json.get("image"));
@@ -70,10 +83,18 @@ public class RestControllerMona {
         throw new IllegalArgumentException("Picture could not be updated");
     }
 
-    @DeleteMapping(value = "/monas/{pinId}/")
-    public void deleteMonaByPinId (@PathVariable("pinId") Long id) {
-        pinRepo.deleteById(id);
-        versionRepo.deletePin(id);
+    @DeleteMapping(value = "/api/monas/{pinId}")
+    public void deleteMonaByPinId (@PathVariable("pinId") Long id, @RequestParam() String username, @RequestHeader Map<String, String> headers) throws Exception {
+        if(headers.containsKey(Config.API_KEY_AUTH_HEADER_NAME_ADMIN) &&                                        //request header has admin API Key
+                headers.get(Config.API_KEY_AUTH_HEADER_NAME_ADMIN).equals(principalRequestValueAdmin) ||        //check if admin API key is correct
+                userRepo.getMappedPins(username).contains(pinRepo.findByPinId(id))){                            //user is the creator of this pin
+
+            pinRepo.deleteById(id);
+            versionRepo.deletePin(id);
+        } else {
+            throw new Exception("Access denied. Only admins and the user, who created this pin are able to delete it");
+        }
+
     }
 
     private boolean addPin(byte[] image, double latitude, double longitude, String username, Long typeId, Date date) {
